emulation: simplify directory structure
emulation._emulate.microphysics was pretty awkward
